api/requirements.txt
fastapi
google-genai
python-multipart
uvicorn
python-dotenv
